requirements.txt
-e .

comet-ml==3.35.5
openai[embeddings]==1.6.0
pinecone-client==2.2.4
promptlayer==0.4.0
python-dotenv==1.0.0
rich==13.3.5
tiktoken==0.5.0
llama-index==0.9.26
alembic==1.13.1
python-json-logger==2.0.7
ddtrace==2.4.0
sentence-transformers==2.2.2
voyageai==0.1.6
